+++ src/net/sourceforge/kolmafia/swingui/menu/LootHermitMenuItem.java
@@ -89,7 +89,7 @@ public class LootHermitMenuItem
 			}
 
 			Integer value = InputFieldUtilities.getQuantity( "How many " + selectedValue.getName() + " to get?\n" + message, maximumValue, 1 );
-			int tradeCount = ( value == null ) ? 0 : value;
+			int tradeCount = ( value == null ) ? 0 : value.intValue();
 
 			if ( tradeCount == 0 )
 			{
